Update domain_name.rst
Just a little miss leading, you would think this is on the page you were linked to, but its back under your name again. I honestly think this needs to be looked at by a developer. Its weird that the just didn't list the "manage databases" in the "your documents" list.